WIP hook createElement to attach {N} wisget to node.
These keeps a reference between the two to ease syncing changes between them. BEcause of this change, the maping of childNodes has to move to be response of MutationObserver changes, rather than on returning the initial wrapper. This should also solve the problem of keeping them insync as lifecycle methods are fired.